requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==django-1.8rc1
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6
static3==0.5.1
djangorestframework==3.1.1
coreapi
